move generic js plotting utils to separate file
